+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/_constants.py
@@ -327,10 +327,4 @@ _DEFAULT_AAD_SCOPE = "https://monitor.azure.com//.default"
 # Default message for messages(MessageData) with empty body
 _DEFAULT_LOG_MESSAGE = "n/a"
 
-# Minimum Severity Level for logs
-_MINIMUM_SEVERITY_LEVEL = "MINIMUM_SEVERITY_LEVEL"
-
-# Trace Based Sampling flag for logs
-_TRACE_BASED_SAMPLING = "TRACE_BASED_SAMPLING"
-
 # cSpell:disable