+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/export/logs/_exporter.py
@@ -2,7 +2,7 @@
 # Licensed under the MIT License.
 import json
 import logging
-from typing import Optional, Sequence, Any, Union
+from typing import Optional, Sequence, Any
 
 from opentelemetry._logs.severity import SeverityNumber
 from opentelemetry.semconv.attributes.exception_attributes import (
@@ -64,7 +64,7 @@ class AzureMonitorLogExporter(BaseExporter, LogExporter):
         :return: The result of the export.
         :rtype: ~opentelemetry.sdk._logs.export.LogData
         """
-        envelopes = [envelope for log in batch if (envelope := self._log_to_envelope(log)) is not None]
+        envelopes = [self._log_to_envelope(log) for log in batch]
         try:
             result = self._transmit(envelopes)
             self._handle_transmit_from_storage(envelopes, result)
@@ -81,10 +81,8 @@ class AzureMonitorLogExporter(BaseExporter, LogExporter):
         if self.storage:
             self.storage.close()
 
-    def _log_to_envelope(self, log_data: LogData) -> Union[TelemetryItem, None]:
+    def _log_to_envelope(self, log_data: LogData) -> TelemetryItem:
         envelope = _convert_log_to_envelope(log_data)
-        if envelope is None:
-            return None
         envelope.instrumentation_key = self._instrumentation_key
         return envelope
 
@@ -119,7 +117,7 @@ def _log_data_is_event(log_data: LogData) -> bool:
 
 # pylint: disable=protected-access
 # pylint: disable=too-many-statements
-def _convert_log_to_envelope(log_data: LogData) -> Union[TelemetryItem, None]:
+def _convert_log_to_envelope(log_data: LogData) -> TelemetryItem:
     log_record = log_data.log_record
     time_stamp = log_record.timestamp if log_record.timestamp is not None else log_record.observed_timestamp
     envelope = _utils._create_telemetry_item(time_stamp)
@@ -205,12 +203,6 @@ def _convert_log_to_envelope(log_data: LogData) -> Union[TelemetryItem, None]:
         if len(data.message) == 0:
             data.message = _DEFAULT_LOG_MESSAGE
         envelope.data = MonitorBase(base_data=data, base_type="MessageData")
-
-    if _utils._should_drop_logs_for_unsampled_traces(log_record): # cspell:disable-line
-        return None
-
-    if _utils._is_less_than_minimum_severity_level(log_record):
-        return None
 
     return envelope