+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.0.0b45 (Unreleased)
 
 ### Features Added
-- Add `minimum_severity_level` and `trace_based_sampling` logger parameters to filter logs
+- Added new log record processor to support Trace Based Sampling
   ([#43811](https://github.com/Azure/azure-sdk-for-python/pull/43811))
 - Added Operation Name Propagation for Dependencies and Logs
   ([#43588](https://github.com/Azure/azure-sdk-for-python/pull/43588))